requirements.txt
# Requirements pour le système Morningstar

## Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
tensorflow>=2.6.0
python-dotenv>=0.19.0
pyyaml>=6.0
scikit-learn>=1.0.0

## API et Connecteurs
ccxt>=2.0.0  # Pour la connexion aux exchanges
requests>=2.26.0  # Pour les appels API

## Logging et Monitoring
loguru>=0.6.0  # Pour une journalisation avancée
pytest>=7.0.0  # Pour les tests

## Tests et Validation
pytest-mock>=3.7.0  # Pour les mocks dans les tests
pytest-cov>=3.0.0  # Pour le coverage des tests

## Développement
black>=22.0  # Formattage du code
flake8>=4.0.0  # Linting
mypy>=0.910  # Vérification des types

## Documentation
mkdocs>=1.2.0  # Pour la documentation
mkdocs-material>=8.0.0  # Thème pour la doc

## Spécifique au Workflow
tqdm>=4.62.0  # Pour les barres de progression
